Start moving types to a shared location
