+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/Infrastructure/EmbeddedSyntax/EmbeddedSyntaxNodeOrToken.cs
@@ -5,7 +5,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.AspNetCore.Analyzers.RouteEmbeddedLanguage.Infrastructure.EmbeddedSyntax;
+namespace Microsoft.AspNetCore.Analyzers.Infrastructure.EmbeddedSyntax;
 
 internal struct EmbeddedSyntaxNodeOrToken<TSyntaxKind, TSyntaxNode>
     where TSyntaxKind : struct
