+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/Infrastructure/RoutePattern/IRoutePatternNodeVisitor.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.AspNetCore.Analyzers.RouteEmbeddedLanguage.RoutePattern;
-
-namespace Microsoft.AspNetCore.Analyzers.RouteEmbeddedLanguage;
+namespace Microsoft.AspNetCore.Analyzers.Infrastructure.RoutePattern;
 
 internal interface IRoutePatternNodeVisitor
 {
