+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/Infrastructure/VirtualChars/AbstractVirtualCharService.cs
@@ -5,11 +5,11 @@ using System;
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Text;
-using Microsoft.AspNetCore.Analyzers.RouteEmbeddedLanguage.Infrastructure.VirtualChars;
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.VirtualChars;
+namespace Microsoft.AspNetCore.Analyzers.Infrastructure.VirtualChars;
 
 internal abstract partial class AbstractVirtualCharService : IVirtualCharService
 {
