+++ src/Framework/AspNetCoreAnalyzers/src/Analyzers/Infrastructure/VirtualChars/System.Text/UnicodeUtility.cs
